split proxy into a file change code example update readme
