+++ README.md
@@ -33,6 +33,9 @@ Some methods in rpc provider is not enabled, here is some examples:
 3. `aggregate` function in mongoose do not auto cast variables to ObjectId or anything else, so you should take case of these variables and do not use them in the `rpc` provider. (for the reason JSON only accept data)
 
 # Changelog
+## 0.1.8
+* auto detective provider
+
 ## 0.1.7
 * auto convert bind port to number
