+++ lib/limbo.js
@@ -8,6 +8,8 @@
 
   EventEmitter = require('events').EventEmitter;
 
+  Proxy = require('./proxy');
+
   Limbo = (function(_super) {
     __extends(Limbo, _super);
 
@@ -31,23 +33,5 @@
   limbo.Limbo = Limbo;
 
   module.exports = limbo;
-
-  Proxy = (function() {
-    function Proxy(_groupName) {
-      this._groupName = _groupName;
-    }
-
-    Proxy.prototype.connect = function(url, callback) {
-      var Provider, providerName;
-      providerName = /^tcp/.test(url) ? 'rpc' : 'mongo';
-      Provider = require("./providers/" + providerName);
-      this.__proto__ = Provider.prototype;
-      Provider.call(this, this._groupName);
-      return this.connect(url, callback);
-    };
-
-    return Proxy;
-
-  })();
 
 }).call(this);
