+++ lib/proxy.js
@@ -1,0 +1,25 @@
+// Generated by CoffeeScript 1.7.1
+(function() {
+  var Proxy;
+
+  Proxy = (function() {
+    function Proxy(_groupName) {
+      this._groupName = _groupName;
+    }
+
+    Proxy.prototype.connect = function(url, callback) {
+      var Provider, providerName;
+      providerName = /^tcp/.test(url) ? 'rpc' : 'mongo';
+      Provider = require("./providers/" + providerName);
+      this.__proto__ = Provider.prototype;
+      Provider.call(this, this._groupName);
+      return this.connect(url, callback);
+    };
+
+    return Proxy;
+
+  })();
+
+  module.exports = Proxy;
+
+}).call(this);
