+++ src/limbo.coffee
@@ -1,5 +1,6 @@
 server = require './server'
 {EventEmitter} = require 'events'
+Proxy = require './proxy'
 
 class Limbo extends EventEmitter
 
@@ -19,13 +20,3 @@ limbo = new Limbo
 limbo.Limbo = Limbo
 
 module.exports = limbo
-
-class Proxy
-  constructor: (@_groupName) ->
-
-  connect: (url, callback) ->
-    providerName = if /^tcp/.test(url) then 'rpc' else 'mongo'
-    Provider = require "./providers/#{providerName}"
-    @__proto__ = Provider.prototype
-    Provider.call(@, @_groupName)
-    @connect(url, callback)
